#76: re-organize search code so it is not all in one file
