+++ assets/js/search/publications/query.js
